workflows/msst/pyproject.toml
[project]
name = "voicelab-msst"
version = "0.1.0"
description = "VoiceLab workflow: MSST (instrumental/vocals/karaoke/dereverb/denoise chain)"
requires-python = ">=3.10,<3.11"
dependencies = [
  # Core
  "torch==2.4.0",
  "numpy==1.26.4",
  "scipy==1.13.1",

  # Audio IO + DSP
  "librosa==0.10.2",
  "soundfile==0.12.1",
  "pydub>=0.25.1",

  # Config + utils (required by upstream MSST code)
  "PyYAML>=6.0.2",
  "ml-collections>=0.1.1",
  "omegaconf==2.3.0",
  "tqdm>=4.66.0",
  "packaging>=24.0",

  # mel_band_roformer deps
  "einops>=0.8.0",
  "rotary-embedding-torch>=0.6.4",
  "beartype>=0.18.5",

  # Downloads
  "requests>=2.32.3",

  # Tests
  "pytest>=8.0.0",
]

[tool.uv]
package = false

